feedkeys(..., 'i') sends keys in the wrong order in Vim 7.0/1/2; prefer :normal and fall back to plain feedkeys(..., '')
The current implementation prefers feedkeys(..., 'i') and only falls back to :normal if Vim 7.4 does not have the 'i' flag and if Vim 7.3 supports :normal with count. For Vim 8.0 (and later), and also for versions between 7.0 and 7.3.99, feedkeys(..., 'i') is used. Due to the prepending action of the 'i' flag, keys must be submitted in reverse order; that's why s is submitted before r . cnt.  Vim 7.0...7.3.99 ignore the 'i' flag, though, and append the keys, which are now in the wrong order (this only matters if a register or count is given).

Instead, prefer to use :normal: If no count is specified, all Vim versions can use it, with count, only Vim 7.3.100 or later. Fallback to feedkeys() (without the 'i' flag) for older Vim versions. The synchronous nature of :normal also helps with advanced customizations that repeat a previous action multiple times through the repeat plugin.
Avoid duplicating the conditional by extracting it into s:normal_with_count().